adicionei sistemas de rotas e as organizei, adicionei tela de cadastro de loja incompleta
