+++ markethub-front-end/src/main.tsx
@@ -3,8 +3,7 @@ import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import "./index.css";
 import App from "./App.tsx";
-import Cadastro from "./Tela-Cadastro.tsx";
-import CadastroProduto from "./Cadastro-produto.tsx";
+
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
