+++ markethub-front-end/src/pages/CadastrarLoja.tsx
@@ -1,0 +1,18 @@
+import HeaderMain from "../components/HeaderMain";
+
+const CadastrarLoja = () => {
+    return(
+      <>
+        <HeaderMain />
+        <h1 className = 'font-bold text-[30px] mt-[45px] ml-[95px]'>
+            Cadastre sua Loja no MarketHub
+        </h1>
+        <p className = 'ml-[95px] text-[#4E4E50]'>
+            Crie sua própria loja online de forma simples e rápida. Qualquer
+             pessoa <br /> pode abrir uma loja gratuita no MarketHub e começar a 
+             vender agora mesmo!
+             </p>
+    </>
+    )
+}
+export default CadastrarLoja;
