+++ markethub-front-end/src/pages/Cadastro-produto.tsx
@@ -1,8 +1,7 @@
-import { StrictMode } from "react";
-import Home from "./App.tsx";
-import { Routes, Route, useLocation } from "react-router-dom";
-import Produtoform from "./components/produtoform.tsx";
-import Header from "./components/header";
+
+
+import { useLocation } from "react-router-dom";
+import Produtoform from "../components/produtoform.tsx";
 
 export default function App() {
   const location = useLocation();
