+++ markethub-front-end/src/pages/Home.tsx
@@ -1,0 +1,9 @@
+import HeaderMain from "../components/HeaderMain";
+
+const Home = () => {
+    return(
+        <HeaderMain />
+    )
+}
+
+export default Home;
